// web/src/constant.d.ts
// 类型声明文件
export interface ConstantType {
  PAGE: number;
  TOTAL: number;
  DEFALUT_SITE_CONFIG: string;
  LOG_TASK_ID: string;
  STORE_TOKEN_NAME: string;
  STORE_AUTH_SOURCE_NAME: string;
  STORE_RBAC_AUTHZ_NAME: string;
  STORE_CUSTOM_NAME: string;
  STORE_OPEN_TABS_NAME: string;
  NO_AUTH_URL: string[];
  WAYS_DATA: Array<{
    type: string;
    [key: string]: any;
  }>;
}

export declare const CONSTANT: ConstantType;